Add files via upload
This code performs a network scan on the `192.168.5.0/24`, which we will have to change, subnet by pinging IP addresses from `192.168.5.1` to `192.168.5.253` in parallel using Python's `threading` library. For each IP address, it constructs and executes a ping command, printing the address if it receives a response. This approach speeds up the process by leveraging multi-threading to conduct multiple pings simultaneously.